// src/typings/func.ts
export enum Func {
  GetAdditionalInfoAplicability = "getAdditionalInfoAplicability",
  GetAdditionalInfoArticle = "getAdditionalInfoArticle",
  GetArticlesForKBA = "GetArticlesForKBA",
  GetArticlesForMime = "GetArticlesForMime",
  GetArticlesForType = "GetArticlesForType",
  GetArtIDByReference = "GetArtidByReference",
  GetBrands = "GetBrands",
  GetCriteriesAplicabilityOfArticle = "GetCriteriesAplicabilityOfArticle",
  GetCriteriesOfArticle = "getCriteriesOfArticle",
  GetCrossReferencesForArticle = "GetCrossReferencesForArticle",
  GetImageForArticle = "getImageForArticle",
  GetInfoArticle = "GetInfoArticle",
  GetInfoCar = "getInfoCar",
  GetInfoEngine = "GetInfoEngine",
  GetLanguages = "getLanguages",
  GetListArticlesFromTradeGroup = "GetListArticlesFromTradeGroup",
  GetListArticlesFromTreeCategory = "GetListArticlesFromTreeCategory",
  GetListEnginesForCar = "getListEnginesForCar",
  GetListReplacedNumbersArticle = "GetListReplacedNumbersArticle",
  GetListTradeGroupsForCar = "GetListTradeGroupsForCar",
  GetManufacturers = "getManufacturers",
  GetModels = "getModels",
  GetPriceFromSupplier = "GetPriceFromSupplier",
  GetSearchArticleByNameForCar = "getSearchArticleByNameForCar",
  GetSearchByCrossref = "getSearchByCrossref",
  GetSearchByCrossrefAndBrand = "GetSearchByCrossrefAndBrand",
  GetSearchCarByName = "GetSearchCarByname",
  GetSearchEngineByCode = "GetSearchEngineByCode",
  GetSearchKBA = "getSearchKBA",
  GetSearchMIME = "getSearchMIME",
  GetSpecificationArticle = "GetSpecificationArticle",
  GetSuppliers = "GetSuppliers",
  GetTreeCategories = "GetTreeCategories",
  GetTypes = "getTypes",
  GetTypesForArticle = "GetTypesForArticle"
}